2次提交

+++ posts/views.py
@@ -1,14 +1,19 @@
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from urllib import quote_plus
 
 
 def post_create(request):
-    form = PostForm(request.POST or None)
+    if request.user.is_anonymous():
+        raise Http404()
+    form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully Created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -18,17 +23,31 @@ def post_create(request):
     return render(request, 'post_form.html', context)
 
 
-def post_detail(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_detail(request, slug=None):
+    instance = get_object_or_404(Post, slug=slug)
+    share_string = quote_plus(instance.content)
     context = {
         "title": instance.title,
-        "instance": instance
+        "instance": instance,
+        'share_string': share_string
     }
     return render(request, 'post_detail.html', context)
 
 
 def post_list(request):
-    queryset =Post.objects.all()
+    queryset_list =Post.objects.all()
+    paginator = Paginator(queryset_list, 5) # Show 25 contacts per page
+
+    page = request.GET.get('page')
+    try:
+        queryset = paginator.page(page)
+    except PageNotAnInteger:
+        # If page is not an integer, deliver first page.
+        queryset = paginator.page(1)
+    except EmptyPage:
+        # If page is out of range (e.g. 9999), deliver last page of results.
+        queryset = paginator.page(paginator.num_pages)
+
     context = {
         "object_list": queryset,
         "title": "List"
@@ -36,9 +55,11 @@ def post_list(request):
     return render(request, 'post_list.html', context)
 
 
-def post_update(request, id=None):
-    instance = get_object_or_404(Post, id=id)
-    form = PostForm(request.POST or None, instance=instance)
+def post_update(request, slug=None):
+    if request.user.is_anonymous():
+        raise Http404()
+    instance = get_object_or_404(Post, slug=slug)
+    form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
         instance.save()
@@ -52,8 +73,10 @@ def post_update(request, id=None):
     return render(request, 'post_form.html', context)
 
 
-def post_delete(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):
+    if request.user.is_anonymous():
+        raise Http404()
+    instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully Deleted")
     return redirect("posts:list")
